+++ R/getBinaryData.R
@@ -28,7 +28,7 @@
 #' @importFrom PamBinaries loadPamguardBinaryFile
 #' @export
 #'
-getBinaryData <- function(x, UID, type=c('click', 'whistle', 'cepstrum'), quiet=FALSE, ...) {
+getBinaryData <- function(x, UID, type=c('click', 'whistle', 'cepstrum', 'gpl'), quiet=FALSE, ...) {
     # if(is.AcousticStudy(x)) {
     #     x <- events(x)
     # }
@@ -77,7 +77,8 @@ getBinaryData <- function(x, UID, type=c('click', 'whistle', 'cepstrum'), quiet=
         typeMatch[t] <- switch(type[t],
                             'click' = '^Click_Detector_|^SoundTrap_Click_Detector_',
                             'whistle' = '^WhistlesMoans_',
-                            'cepstrum' = '^WhistlesMoans_'
+                            'cepstrum' = '^WhistlesMoans_',
+                            'gpl' = '^GPL_Detector_'
         )
     }
     typeMatch <- paste0(typeMatch, collapse='|')